Bump and relax dependency version restrictions
We assume some level of SemVer compliance with our dependencies and add
a caret to all version requirements. Any major version bumps will need
to be reviewed by hand and then adjusted in the project file. For all
other cases, a simple `poetry update` should work and suffice to update
the dependencies to the latest supported version.