mpris: Don't log ignored events
The event listener is quite reliable and we really don't need to log
each event from an inactive player when we can instead log their names
when we hear about them for the first time.

Spotify is known to send at least 5 metadata updates whenever a track
changes, for example.