Relax requirements on pillow
pillow 10 is required for Python 3.12 (and also still supports 3.8).

By using an open-ended requirement, we won't easily run into this again
on the next Python (and pillow) update, when 3.8 will likely be dropped
and there are barely any relevant API changes for us. If there are,
it's easier to fix them rather than being over-specific in our
requirements.

Fixes #728