Replace double-fork with a polling thread
The double fork method has some issues, notably that it is quite fragile
and that any errors that occur while spawning the process, e.g. that no
executable can be found with the requested name, are swallowed.

Instead, we use the saner workaround with a polling daemon thread that
simply polls all our spawned processes every 5 seconds to reap them if
they have exited. This has all the functionality that we desire with
the cost of just a simple background thread that does nothing most of
the time.

Fixes #525
Fixes #508