call new signal "finish" when player is closed after an "update"
Created a new signal called "finish" which is called whenever a player
is closed (tested with mpris) whenever an update is due. This means if
an episode is just completed and it is needed to be updated, this signal
is called when that player is closed.

Intended to be used as a "episode completed" indicator for various uses
like autoplaying next episode