mpris: Fix tracker state updating
Updates for the current player that did not result in a new file being
played would cause the tracker state to be set to `None`, resulting in
the tracker to not update the currently watching episode when the timer
expires.

Fixes a regression from e922f456f1ea30b5328485567a55562c02c2c9bd.