data: Safely store data on disk (attempt 2)
Write to a temporary file first and only then remove the original file
and rename the temporary to the intended file name. This ensures that
we do not write partial files when the disk is full, for example.

Use a different strategy for symlinked files because we need to write to
those directly.

Additionally, streamline encoding to always use UTF-8.